Move DictionaryContentsComparer out of top-level
