+++ sdk/dotnet/Pulumi.Automation/ProjectTemplate.cs
@@ -2,6 +2,7 @@
 
 using System;
 using System.Collections.Generic;
+using Pulumi.Automation.Collections;
 
 namespace Pulumi.Automation
 {
